Update Protocol 12.51

+++ engine/src/protocolgamebase.cpp
@@ -1269,7 +1269,7 @@ void ProtocolGameBase::sendPreyData(uint8_t preySlotId)
 		}
 	}
 
-	msg.add<uint16_t>(player->getFreeRerollTime(preySlotId));
+	msg.add<uint32_t>(player->getFreeRerollTime(preySlotId));
 	if (version >= 1190) {
 		msg.addByte(0x00); //preyWildCards
 	}
